receive-job-inquiry-mail/requirements.txt
flask==3.1.2
gunicorn==25.1.0
httpx==0.28.1
google-cloud-secret-manager==2.26.0
google-cloud-logging==3.13.0
google-cloud-firestore==2.23.0
google-api-python-client==2.190.0
google-auth==2.40.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.1
pytz==2025.2
